:recycle: Refactor: 장바구니, 적립유무 화면 병합
