+++ src/pages/TouchOrder.jsx
@@ -43,7 +43,7 @@ function TouchOrderContent() {
 
   /* 내부 동작 함수 선언식 */
   function handleCartClick() {
-    navigate("/order/touch/cart");
+    navigate("/order/cart");
   }
 
   function handleAddToCart(product, quantity) {
